Add metrics exporter (statsd)
- Refactoring.
- Fix compile errors when no `metrics` feature is specified.
- Temporary print the stats from Moka v0.11.1 (unreleased).
