+++ Cargo.toml
@@ -56,9 +56,9 @@ arc-swap = "1.6.0"
 [dependencies.moka011]
 package = "moka"
 optional = true
-version = "0.11.0"
-# git = "https://github.com/moka-rs/moka"
-# branch = "master"
+# version = "0.11.1"
+git = "https://github.com/moka-rs/moka"
+branch = "cache-stats-v0.11"
 # path = "../moka"
 features = ["future"]
