+++ src/cache/moka_driver/sync_segmented.rs
@@ -162,6 +162,11 @@ impl<I> MokaSegmentedCache<I> {
                 eviction_counters = None;
             }
 
+            #[cfg(not(any(feature = "moka-v09", feature = "moka-v010")))]
+            {
+                builder = builder.enable_stats();
+            }
+
             cache = builder.build_with_hasher(DefaultHasher::default());
         }
 
@@ -246,6 +251,11 @@ impl<I: GetOrInsertOnce> CacheDriver<TraceEntry> for MokaSegmentedCache<I> {
 
     fn eviction_counters(&self) -> Option<Arc<EvictionCounters>> {
         self.eviction_counters.as_ref().map(Arc::clone)
+    }
+
+    #[cfg(not(any(feature = "moka-v08", feature = "moka-v09", feature = "moka-v010")))]
+    fn cache_stats(&self) -> Option<crate::moka::stats::CacheStats> {
+        Some(self.cache.stats())
     }
 }
