+++ src/metrics_exporter/disabled_exporter.rs
@@ -1,0 +1,3 @@
+pub(crate) async fn init(_endpoint: &str) {}
+
+pub(crate) fn shutdown() {}
